add FindCommodity domain and FindCommodityFilter objects convert all primatives to Object Wrappers create find_commodity_view convert demand/supply to Long
TODO: distance mapping and default/non-null argument handling, findCommodityFilterEntity?, Rename FindCommodity mappings to LocateCommodity?, sort out mybatis supply/demand if statements

+++ trade-module/src/main/java/io/edpn/backend/trade/application/dto/CommodityMarketInfoResponse.java
@@ -17,8 +17,8 @@ public class CommodityMarketInfoResponse {
     Double minMeanPrice;
     Double maxMeanPrice;
     Double averageMeanPrice;
-    Integer totalStock;
-    Integer totalDemand;
+    Long totalStock;
+    Long totalDemand;
     Integer totalStations;
     Integer stationsWithBuyPrice;
     Integer stationsWithSellPrice;
